\name{getModelFrame}
\alias{getModelFrame}
\alias{getModelFrame,Gmp-method}
\title{Get Model Frame Gmp Object}
\description{Retrieve the model frame from a \code{Gmp} object.}
\usage{
getModelFrame(object)
}
\arguments{
  \item{object}{A Gmp object, created by \link{gmpCreate} or \link{gmp}.}
}
\value{
  A data.frame with various attributes.
}
