\name{gmpCoef}
\alias{gmpCoef}
\title{Get Parameter Estimates from Original Fit}
\description{
  Retrieve the parameter estimates from a regression fit to the data
  with the original labeling.
}
\usage{
gmpCoef(x)
}
\arguments{
  \item{x}{A \code{Gmp} object}
}
\value{
  A vector (in the normal case), or a matrix (for multinomial data).
}
