\name{gmpm-package}
\alias{gmpm-package}
\alias{gmpm}
\docType{package}
\title{
  Generalized Multilevel Permutation Models (GMPMs)
}
\description{
  Functions for fitting GMPMs.  Documentation currently under
  construction (sorry).
}
\details{
\tabular{ll}{
Package: \tab gmpm\cr
Type: \tab Package\cr
Version: \tab 1.0-2\cr
Date: \tab 2009-08-27\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
Depends: \tab methods\cr
}
}
\author{
  Dale J. Barr <dale.barr@ucr.edu>
}
\references{
Barr, D. J. (in preparation).  Generalized Multilevel
Permutation Models.
}

%~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
%~~ the R documentation directory ~~

\keyword{ package }
\keyword{htest}
\keyword{robust}
\keyword{design}
\keyword{ts}
\keyword{nonparametric}
\keyword{multivariate}
\seealso{
  \code{\link{gmp}}
}
