\name{gmpmCoef}
\alias{gmpmCoef}
\alias{gmpmCoef,GMPM-method}
\title{Get Parameter Estimates from Original Fit}
\description{
  Retrieve the parameter estimates from a regression fit to the data
  with the original labeling.
}
\usage{
gmpmCoef(x)
}
\arguments{
  \item{x}{A \code{GMPM} object}
}
\value{
  A vector (in the normal case), or a matrix (for multinomial data).
}
