\name{getModelFrame}
\alias{getModelFrame}
\alias{getModelFrame,GMPM-method}
\title{Get Model Frame GMPM Object}
\description{Retrieve the model frame from a \code{GMPM} object.}
\usage{
getModelFrame(object)
}
\arguments{
  \item{object}{A GMPM object, created by \link{gmpmCreate} or \link{gmpm}.}
}
\value{
  A data.frame with various attributes.
}
